feat(browser): author-2d 整本模式 — 浏览器一键 新闻→10-20页+PPTX; news 链进 eval 语料
Sprint 32, (a)+(b) per user.

(b) Browser full-deck mode. The pipeline core moved from the bake script
into src so both ends share ONE implementation (never fork):
- src/present/scaffolds/lift-slot-llm.js — Stage 2 system+user prompt
  (Rules 0-20) + parser, extracted verbatim (dry-run prompt char counts
  identical before/after)
- src/present/news/expand-core.js — Stage -1 expansion (CLI wrapper stays
  at scripts/expand-news.mjs)
- src/present/news/full-deck.js — browser orchestrator: expand → map →
  orphan rescue → page floor → per-slot lift with progress callbacks
- mapper-llm + lift get the anthropic-dangerous-direct-browser-access
  header (BYOK direct calls, same as text-to-ir)

author-2d gains a 快速/整本 mode select. Verified live: tech-chip article
→ 14 pages in 120s, 0 slot failures, PPTX downloaded from the page.

(a) eval-corpus supports article entries baked through news-to-deck.mjs
— the full news chain now regresses with the corpus. New 13-deck
baseline: 97.2 composite / facts 95.7% / entities 95.7%; both news
entries score 100 (14/14 slots, facts 100%).

Tests: 126/126.
