+++ sdf-js/examples/scaffold-pipeline/eval-news-econ-slidedata.json
@@ -1,0 +1,131 @@
+[
+  {
+    "title": "全球经济2026年中展望",
+    "body": ["摘录自国际机构与财经媒体报道", "覆盖全球增长预测、区域表现、货币政策与主要风险"]
+  },
+  {
+    "title": "核心议题概览",
+    "body": [
+      "三大国际机构对2026-2027年全球经济增速的预测差异",
+      "主要经济体与地区的增长分化格局",
+      "美联储等央行的货币政策调整路径",
+      "中东冲突、科技投资、就业与金融市场的四类风险评估"
+    ]
+  },
+  {
+    "title": "国际货币基金组织增长预测",
+    "body": [
+      "最新《世界经济展望》将2026年全球增长预期定为3.3%",
+      "2027年全球增长预期为3.2%",
+      "显示中期全球经济温和扩张趋势"
+    ]
+  },
+  {
+    "title": "经合组织中期展望",
+    "body": [
+      "中期报告认为全球GDP增速将大致稳定在2.9%",
+      "预测较IMF更为保守",
+      "反映发达经济体结构性增长约束"
+    ]
+  },
+  {
+    "title": "世界银行保守预测与警告",
+    "body": [
+      "世界银行的预测更为保守,为2.5%",
+      "警告新兴市场和发展中经济体将面临疫情以来最弱的人均收入增长",
+      "突显发展中国家复苏面临的结构性挑战"
+    ]
+  },
+  {
+    "title": "中国经济增长预期",
+    "body": [
+      "中国2026年实际GDP预计增长5%",
+      "增速在主要经济体中保持领先",
+      "为全球增长提供重要动力来源"
+    ]
+  },
+  {
+    "title": "欧元区温和增长",
+    "body": [
+      "欧元区增长预计维持在1.1%的温和水平",
+      "反映欧洲经济结构调整与能源转型压力",
+      "增速显著低于全球平均水平"
+    ]
+  },
+  {
+    "title": "低收入国家增长潜力",
+    "body": [
+      "低收入国家增速有望达到5.4%",
+      "成为全球增长最快的经济体群组",
+      "但世界银行警告其人均收入增长仍面临压力"
+    ]
+  },
+  {
+    "title": "通胀回落与降息空间",
+    "body": [
+      "多数主要经济体通胀持续回落",
+      "美联储、英国央行与欧洲央行均有降息空间",
+      "为宽松货币政策转向创造条件"
+    ]
+  },
+  {
+    "title": "美联储降息路径预期",
+    "body": [
+      "市场预计美联储将在4月前继续降息",
+      "前提是美国就业增长放缓且核心通胀涨幅温和",
+      "当政策利率降至3%-3.25%区间后,预计将进入较长时间的观望期",
+      "降息周期节奏取决于劳动力市场与通胀数据"
+    ]
+  },
+  {
+    "title": "中东冲突能源价格冲击",
+    "body": [
+      "中东冲突已引发能源价格大幅上涨、通胀重新抬头以及货币政策收紧的预期",
+      "霍尔木兹海峡航运中断和部分能源基础设施关闭",
+      "导致能源价格激增,并扰乱了能源及其他重要大宗商品的全球供应",
+      "其发生概率高、冲击程度深"
+    ]
+  },
+  {
+    "title": "科技投资上行支撑",
+    "body": [
+      "科技相关投资与生产的强劲势头构成上行支撑",
+      "但其持续性存在不确定性",
+      "可能成为全球经济增长的积极因素"
+    ]
+  },
+  {
+    "title": "就业市场下行风险",
+    "body": [
+      "就业市场超预期走弱属于中等概率、中等影响的风险",
+      "可能影响消费支出与经济增长动能",
+      "是货币政策决策的关键观察指标"
+    ]
+  },
+  {
+    "title": "金融市场估值风险",
+    "body": [
+      "金融市场估值回调风险概率较低但影响可能较大",
+      "反映当前资产价格水平存在的潜在脆弱性",
+      "一旦触发可能对全球金融稳定造成显著冲击"
+    ]
+  },
+  {
+    "title": "风险矩阵总览",
+    "body": [
+      "中东冲突能源冲击:高概率、高影响",
+      "科技投资势头:上行支撑但持续性不确定",
+      "就业市场走弱:中等概率、中等影响",
+      "金融估值回调:低概率、高影响"
+    ]
+  },
+  {
+    "title": "2026年展望总结",
+    "body": [
+      "全球经济预计温和增长,三大机构预测在2.5%-3.3%区间",
+      "区域分化显著:中国与低收入国家增速领先,欧元区相对疲弱",
+      "货币政策转向宽松,但降息路径取决于通胀与就业数据",
+      "地缘政治与能源供应构成最主要下行风险,科技投资提供上行可能"
+    ]
+  }
+]
