+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/eval.json
@@ -1,0 +1,59 @@
+{
+  "deckDir": "/Users/hexiaoyang/Documents/sdf-main/sdf-js/examples/scaffold-pipeline/eval-news-econ",
+  "deckName": "eval-news-econ",
+  "structure": {
+    "slots_total": 14,
+    "slots_baked": 14,
+    "slots_empty": 0,
+    "slots_errored": 0,
+    "fill_rate": 1,
+    "pick_method": "forced",
+    "scaffold_id": "news-briefing",
+    "theme_id": "editorial-navy"
+  },
+  "atomQuality": {
+    "atom_instances": 19,
+    "atom_types_distinct": 15,
+    "variety_ratio": 0.7894736842105263,
+    "unknown_atom_count": 0,
+    "unknown_atom_types": [],
+    "arg_violations": 0,
+    "arg_violations_top_offenders": []
+  },
+  "threeDReadiness": {
+    "twin_coverage": 1,
+    "untwinned_types": [],
+    "lift_success": 14,
+    "lift_attempted": 14,
+    "lift_errors": [],
+    "lifted_subject_rate": 1
+  },
+  "textBudget": {
+    "chars_per_slot": 82.28571428571429,
+    "max_slot_chars": 130
+  },
+  "fidelity": {
+    "numbers_total": 11,
+    "numbers_found": 11,
+    "number_recall": 1,
+    "missing_numbers": [],
+    "entities_total": 6,
+    "entities_found": 5,
+    "entity_recall": 0.8333333333333334,
+    "missing_entities": [
+      "但世界银行"
+    ]
+  },
+  "score": {
+    "total": 100,
+    "breakdown": {
+      "structure": 25,
+      "variety": 15,
+      "hallucination": 15,
+      "twin": 20,
+      "lift": 15,
+      "text": 10
+    }
+  },
+  "generatedAt": "2026-07-07T01:20:05.708Z"
+}
