+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-00-cover.json
@@ -1,0 +1,51 @@
+{
+  "slotIdx": 0,
+  "slotName": "cover",
+  "slotTitle": "Headline",
+  "slotPurpose": "Article headline + source and date context",
+  "sourceSlideIdx": 0,
+  "sourceSlideTitle": "全球经济2026年中展望",
+  "sceneData": {
+    "name": "news-briefing/cover",
+    "layout": "cover",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 720,
+        "args": {
+          "title": "全球经济2026年中展望",
+          "subtitle": "摘录自国际机构与财经媒体报道",
+          "author": "覆盖全球增长预测、区域表现、货币政策与主要风险",
+          "style": "gradient"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "cover"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4404,
+      "cache_creation_input_tokens": 13008,
+      "cache_read_input_tokens": 0,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 13008,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 177,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.064647,
+    "elapsedSec": 5.1,
+    "generatedAt": "2026-07-07T01:15:51.595Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
