+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-01-agenda.json
@@ -1,0 +1,78 @@
+{
+  "slotIdx": 1,
+  "slotName": "agenda",
+  "slotTitle": "Overview",
+  "slotPurpose": "One line per major theme of the briefing",
+  "sourceSlideIdx": 1,
+  "sourceSlideTitle": "核心议题概览",
+  "sceneData": {
+    "name": "news-briefing/agenda",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "核心议题概览",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "agenda-list",
+        "x": 80,
+        "y": 160,
+        "w": 1120,
+        "h": 520,
+        "args": {
+          "items": [
+            {
+              "label": "三大国际机构2026-2027预测差异",
+              "sublabel": "经合组织中期展望稳定在2.9%，较IMF更为保守"
+            },
+            {
+              "label": "主要经济体增长分化格局",
+              "sublabel": "欧元区1.1%温和增长，低收入国家5.4%增速最快"
+            },
+            {
+              "label": "美联储等央行货币政策调整路径"
+            },
+            {
+              "label": "四类风险评估",
+              "sublabel": "中东冲突、科技投资、就业市场下行、金融市场估值回调"
+            }
+          ],
+          "numbered": true,
+          "style": "showcase"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "agenda-list",
+      "number-list"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 5064,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 398,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0250644,
+    "elapsedSec": 9.3,
+    "generatedAt": "2026-07-07T01:16:00.945Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
