+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-02-key-figures.json
@@ -1,0 +1,73 @@
+{
+  "slotIdx": 2,
+  "slotName": "key-figures",
+  "slotTitle": "Key Figures",
+  "slotPurpose": "The headline numbers at a glance",
+  "sourceSlideIdx": 2,
+  "sourceSlideTitle": "国际货币基金组织增长预测",
+  "sceneData": {
+    "name": "news-briefing/key-figures",
+    "layout": "grid",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "关键数据",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "stat-grid-large",
+        "x": 80,
+        "y": 200,
+        "w": 1120,
+        "h": 420,
+        "args": {
+          "title": "国际货币基金组织增长预测",
+          "stats": [
+            {
+              "value": "3.3%",
+              "label": "2026年全球增长预期",
+              "trend": "《世界经济展望》"
+            },
+            {
+              "value": "3.2%",
+              "label": "2027年全球增长预期",
+              "trend": "中期温和扩张"
+            }
+          ]
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "stat-grid-large",
+      "kpi-card",
+      "dashboard-multi-kpi-composite"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4457,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 301,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0217884,
+    "elapsedSec": 6.2,
+    "generatedAt": "2026-07-07T01:16:07.103Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
