+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-03-theme-1-lead.json
@@ -1,0 +1,80 @@
+{
+  "slotIdx": 3,
+  "slotName": "theme-1-lead",
+  "slotTitle": "Theme 1 — Lead",
+  "slotPurpose": "First major theme: the core facts and figures",
+  "sourceSlideIdx": 2,
+  "sourceSlideTitle": "国际货币基金组织增长预测",
+  "sceneData": {
+    "name": "news-briefing/theme-1-lead",
+    "layout": "grid",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "国际货币基金组织增长预测",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 80,
+        "y": 200,
+        "w": 480,
+        "h": 400,
+        "args": {
+          "value": "3.3%",
+          "label": "2026年全球增长",
+          "sublabel": "《世界经济展望》",
+          "icon": "chart-line-up",
+          "style": "dark"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 720,
+        "y": 200,
+        "w": 480,
+        "h": 400,
+        "args": {
+          "value": "3.2%",
+          "label": "2027年全球增长",
+          "sublabel": "中期温和扩张",
+          "icon": "globe",
+          "style": "light"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "bar",
+      "column",
+      "stat-banner",
+      "kpi-card"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4462,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 361,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0227034,
+    "elapsedSec": 6.1,
+    "generatedAt": "2026-07-07T01:16:13.220Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
