+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-04-theme-1-detail.json
@@ -1,0 +1,112 @@
+{
+  "slotIdx": 4,
+  "slotName": "theme-1-detail",
+  "slotTitle": "Theme 1 — Detail",
+  "slotPurpose": "First theme drill-down: regional / per-entity breakdown",
+  "sourceSlideIdx": 5,
+  "sourceSlideTitle": "中国经济增长预期",
+  "sceneData": {
+    "name": "news-briefing/theme-1-detail",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "中国经济增长预期",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 80,
+        "y": 200,
+        "w": 340,
+        "h": 420,
+        "args": {
+          "value": "5%",
+          "label": "2026年实际GDP增长",
+          "sublabel": "中国",
+          "icon": "chart-line-up",
+          "style": "dark",
+          "trend": "up",
+          "trendValue": "领先"
+        }
+      },
+      {
+        "type": "stat-with-icon",
+        "x": 480,
+        "y": 200,
+        "w": 340,
+        "h": 200,
+        "args": {
+          "value": "领先",
+          "label": "主要经济体中",
+          "sublabel": "增速排名",
+          "icon": "trophy",
+          "iconColor": "rgb(38, 70, 130)"
+        }
+      },
+      {
+        "type": "stat-with-icon",
+        "x": 480,
+        "y": 420,
+        "w": 340,
+        "h": 200,
+        "args": {
+          "value": "重要",
+          "label": "全球增长动力",
+          "sublabel": "贡献来源",
+          "icon": "globe-hemisphere-west",
+          "iconColor": "rgb(38, 70, 130)"
+        }
+      },
+      {
+        "type": "icon-badge",
+        "x": 880,
+        "y": 260,
+        "w": 280,
+        "h": 300,
+        "args": {
+          "name": "chart-bar",
+          "label": "经济增长",
+          "color": [
+            38,
+            70,
+            130
+          ]
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "bar",
+      "pie",
+      "segmented-bar",
+      "isotype-stat-comparison"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4448,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 588,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.026066399999999997,
+    "elapsedSec": 9.3,
+    "generatedAt": "2026-07-07T01:16:22.486Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
