+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-05-theme-2-lead.json
@@ -1,0 +1,76 @@
+{
+  "slotIdx": 5,
+  "slotName": "theme-2-lead",
+  "slotTitle": "Theme 2 — Lead",
+  "slotPurpose": "Second major theme: the core facts",
+  "sourceSlideIdx": 8,
+  "sourceSlideTitle": "通胀回落与降息空间",
+  "sceneData": {
+    "name": "news-briefing/theme-2-lead",
+    "layout": "stage",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "通胀回落与降息空间",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "process-arrows",
+        "x": 80,
+        "y": 200,
+        "w": 1120,
+        "h": 420,
+        "args": {
+          "steps": [
+            {
+              "label": "通胀持续回落",
+              "sublabel": "多数主要经济体"
+            },
+            {
+              "label": "央行降息空间",
+              "sublabel": "美联储、英国央行与欧洲央行"
+            },
+            {
+              "label": "宽松货币政策",
+              "sublabel": "转向条件创造"
+            }
+          ],
+          "color": "gradient"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "process-arrows",
+      "timeline",
+      "stat-banner",
+      "kpi-card"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4453,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 312,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0219414,
+    "elapsedSec": 5.9,
+    "generatedAt": "2026-07-07T01:16:28.342Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
