+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-06-theme-2-detail.json
@@ -1,0 +1,103 @@
+{
+  "slotIdx": 6,
+  "slotName": "theme-2-detail",
+  "slotTitle": "Theme 2 — Detail",
+  "slotPurpose": "Second theme drill-down: sequence, conditions, thresholds",
+  "sourceSlideIdx": 9,
+  "sourceSlideTitle": "美联储降息路径预期",
+  "sceneData": {
+    "name": "news-briefing/theme-2-detail",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "美联储降息路径预期",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "timeline",
+        "x": 80,
+        "y": 180,
+        "w": 1120,
+        "h": 280,
+        "args": {
+          "events": [
+            {
+              "date": "4月前",
+              "label": "继续降息",
+              "sublabel": "前提: 就业增长放缓 + 核心通胀温和"
+            },
+            {
+              "date": "3%-3.25%",
+              "label": "进入观望期",
+              "sublabel": "政策利率降至目标区间"
+            },
+            {
+              "date": "观望期",
+              "label": "较长时间暂停",
+              "sublabel": "节奏取决于劳动力市场与通胀数据"
+            }
+          ],
+          "axisLabel": "降息周期时间线"
+        }
+      },
+      {
+        "type": "bullet-list",
+        "x": 80,
+        "y": 500,
+        "w": 1120,
+        "h": 180,
+        "args": {
+          "title": "关键前提条件",
+          "items": [
+            {
+              "icon": "users",
+              "label": "美国就业增长放缓"
+            },
+            {
+              "icon": "chart-line",
+              "label": "核心通胀涨幅温和"
+            },
+            {
+              "icon": "gauge",
+              "label": "政策利率降至3%-3.25%区间后观望"
+            }
+          ]
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "timeline",
+      "funnel-with-conversion",
+      "bullet-list",
+      "flow-chart"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4508,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 549,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.025661399999999997,
+    "elapsedSec": 8.6,
+    "generatedAt": "2026-07-07T01:16:36.961Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
