+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-07-theme-3-lead.json
@@ -1,0 +1,98 @@
+{
+  "slotIdx": 7,
+  "slotName": "theme-3-lead",
+  "slotTitle": "Theme 3 — Lead",
+  "slotPurpose": "Third major theme: the core facts",
+  "sourceSlideIdx": 10,
+  "sourceSlideTitle": "中东冲突能源价格冲击",
+  "sceneData": {
+    "name": "news-briefing/theme-3-lead",
+    "layout": "stage",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "能源价格冲击",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "icon-grid",
+        "x": 80,
+        "y": 180,
+        "w": 1120,
+        "h": 480,
+        "args": {
+          "title": "中东冲突能源价格冲击",
+          "items": [
+            {
+              "icon": "chart-line-up",
+              "label": "能源价格大幅上涨",
+              "sublabel": "通胀重新抬头"
+            },
+            {
+              "icon": "currency-dollar",
+              "label": "货币政策收紧",
+              "sublabel": "预期强化"
+            },
+            {
+              "icon": "boat",
+              "label": "霍尔木兹海峡",
+              "sublabel": "航运中断"
+            },
+            {
+              "icon": "plugs",
+              "label": "能源基础设施",
+              "sublabel": "部分关闭"
+            },
+            {
+              "icon": "lightning-slash",
+              "label": "能源价格激增",
+              "sublabel": "全球供应扰乱"
+            },
+            {
+              "icon": "warning-diamond",
+              "label": "高概率深冲击",
+              "sublabel": "风险显著"
+            }
+          ],
+          "cols": 3,
+          "iconStyle": "circle",
+          "iconSize": "medium",
+          "colorMode": "brand"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "stat-banner",
+      "kpi-card",
+      "bar",
+      "bullet-list"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4521,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 508,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0250854,
+    "elapsedSec": 8.4,
+    "generatedAt": "2026-07-07T01:16:45.384Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
