+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-10-risk-detail-2.json
@@ -1,0 +1,74 @@
+{
+  "slotIdx": 10,
+  "slotName": "risk-detail-2",
+  "slotTitle": "Secondary Risks",
+  "slotPurpose": "Remaining risks / upside factors",
+  "sourceSlideIdx": 11,
+  "sourceSlideTitle": "科技投资上行支撑",
+  "sceneData": {
+    "name": "news-briefing/risk-detail-2",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "Secondary Risks",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "balance-scale",
+        "x": 80,
+        "y": 180,
+        "w": 1120,
+        "h": 480,
+        "args": {
+          "title": "科技投资上行支撑",
+          "leftLabel": "上行因素",
+          "rightLabel": "不确定性",
+          "leftItems": [
+            "科技投资强劲势头",
+            "生产增长支撑",
+            "全球经济积极因素"
+          ],
+          "rightItems": [
+            "持续性存在不确定性",
+            "投资波动风险"
+          ],
+          "leftAccent": "rgb(38, 70, 130)",
+          "rightAccent": "rgb(110, 95, 75)"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "balance-scale",
+      "icon-grid",
+      "bullet-list"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4440,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 332,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.022202399999999997,
+    "elapsedSec": 6.7,
+    "generatedAt": "2026-07-07T01:17:06.885Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
