+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-11-quote.json
@@ -1,0 +1,63 @@
+{
+  "slotIdx": 11,
+  "slotName": "quote",
+  "slotTitle": "Key Quote",
+  "slotPurpose": "A load-bearing quote or judgment from the article",
+  "sourceSlideIdx": 4,
+  "sourceSlideTitle": "世界银行保守预测与警告",
+  "sceneData": {
+    "name": "news-briefing/quote",
+    "layout": "stage",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "Key Quote",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "pull-quote-banner",
+        "x": 80,
+        "y": 180,
+        "w": 1120,
+        "h": 460,
+        "args": {
+          "quote": "新兴市场和发展中经济体将面临疫情以来最弱的人均收入增长",
+          "author": "世界银行",
+          "attribution": "2025年全球经济展望 · 2.5%增长预测",
+          "bg": "accent"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "quote-pull",
+      "pull-quote-banner"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4453,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 257,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0211164,
+    "elapsedSec": 5.5,
+    "generatedAt": "2026-07-07T01:17:12.342Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
