+++ sdf-js/examples/scaffold-pipeline/eval-news-econ/slots/slot-12-outlook.json
@@ -1,0 +1,79 @@
+{
+  "slotIdx": 12,
+  "slotName": "outlook",
+  "slotTitle": "Outlook",
+  "slotPurpose": "What to watch next — timeline or conditions",
+  "sourceSlideIdx": 9,
+  "sourceSlideTitle": "美联储降息路径预期",
+  "sceneData": {
+    "name": "news-briefing/outlook",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "Outlook",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "timeline",
+        "x": 80,
+        "y": 180,
+        "w": 1120,
+        "h": 480,
+        "args": {
+          "title": "美联储降息路径预期",
+          "events": [
+            {
+              "date": "4月前",
+              "label": "继续降息",
+              "sublabel": "前提：就业增长放缓、核心通胀温和"
+            },
+            {
+              "date": "3%-3.25%",
+              "label": "进入观望期",
+              "sublabel": "政策利率区间"
+            },
+            {
+              "date": "后续",
+              "label": "节奏取决数据",
+              "sublabel": "劳动力市场与通胀指标"
+            }
+          ],
+          "axisLabel": "降息周期"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "timeline",
+      "mountain-path",
+      "bullet-list"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4483,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 355,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0226764,
+    "elapsedSec": 12.3,
+    "generatedAt": "2026-07-07T01:17:24.615Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
