+++ sdf-js/examples/scaffold-pipeline/eval-news-policy-slidedata.json
@@ -1,0 +1,101 @@
+[
+  {
+    "title": "全国新能源汽车下乡政策成效与挑战",
+    "body": ["财经报纸报道摘录", "聚焦县乡市场新能源汽车推广现状", "数据截至今年前五个月"]
+  },
+  {
+    "title": "核心议题概览",
+    "body": [
+      "下乡车型销量增长与市场渗透现状",
+      "充电桩等配套设施建设进展",
+      "主要车企与产品市场格局",
+      "农村电网、残值体系等关键挑战",
+      "政策退坡对未来增速的影响"
+    ]
+  },
+  {
+    "title": "下乡车型销量高速增长",
+    "body": [
+      "今年前五个月新能源汽车下乡车型销量达到187万辆",
+      "同比增长61%",
+      "占全国新能源汽车总销量的34%"
+    ]
+  },
+  {
+    "title": "县乡市场渗透率提升明显",
+    "body": [
+      "县乡市场渗透率从去年同期的17%提升至28%",
+      "仍显著低于一二线城市的54%",
+      "说明下沉市场仍有较大增长空间"
+    ]
+  },
+  {
+    "title": "充电桩建设成最大短板",
+    "body": ["截至5月底全国县域公共充电桩保有量约98万台", "仅占全国总量的21%", "平均每县不足350台"]
+  },
+  {
+    "title": "充电设施覆盖目标与时间表",
+    "body": [
+      "国家能源局要求2026年底前实现'县县全覆盖、乡乡有站点'",
+      "为县域充电网络建设设定明确时间节点",
+      "涉及全国数千个县级行政区域"
+    ]
+  },
+  {
+    "title": "充电基础设施财政支持力度",
+    "body": [
+      "中央财政安排专项补贴120亿元",
+      "地方配套资金约200亿元",
+      "合计投入超过300亿元用于基础设施建设"
+    ]
+  },
+  {
+    "title": "比亚迪、吉利、长安领跑下乡市场",
+    "body": [
+      "比亚迪、吉利、长安三家车企合计占据下乡车型销量的58%",
+      "头部车企在县乡市场形成集中优势",
+      "反映品牌认知度与渠道下沉能力的重要性"
+    ]
+  },
+  {
+    "title": "五菱宏光MINI等微型车贡献突出",
+    "body": [
+      "五菱宏光MINI等微型电动车单一车型贡献超过15%",
+      "小型化、低价化产品契合农村消费需求",
+      "单车型占比显示细分市场集中度"
+    ]
+  },
+  {
+    "title": "5万至10万元价格带主导市场",
+    "body": [
+      "5万至10万元车型占比达63%",
+      "反映县乡消费者价格敏感特征",
+      "经济型车型是下乡政策主要受益品类"
+    ]
+  },
+  {
+    "title": "挑战一：农村电网改造滞后",
+    "body": ["部分乡镇变压器容量不足", "夏季用电高峰期充电功率受限", "该问题发生概率高、影响中等"]
+  },
+  {
+    "title": "挑战二：二手车残值体系缺失",
+    "body": [
+      "二手车残值体系缺失导致置换意愿低",
+      "属于中等概率、深远影响的问题",
+      "影响新能源汽车在农村市场的长期普及"
+    ]
+  },
+  {
+    "title": "挑战三：补贴退坡风险",
+    "body": ["若购置补贴按期退坡", "明年销量增速可能回落至30%以下", "政策依赖度仍然较高"]
+  },
+  {
+    "title": "前景与展望",
+    "body": [
+      "县乡市场成为新能源汽车增长新引擎",
+      "基础设施、电网配套、残值体系建设亟待加速",
+      "政策延续性与市场化转型需平衡推进",
+      "渗透率提升空间仍然广阔"
+    ]
+  }
+]
