+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/eval.json
@@ -1,0 +1,57 @@
+{
+  "deckDir": "/Users/hexiaoyang/Documents/sdf-main/sdf-js/examples/scaffold-pipeline/eval-news-policy",
+  "deckName": "eval-news-policy",
+  "structure": {
+    "slots_total": 14,
+    "slots_baked": 14,
+    "slots_empty": 0,
+    "slots_errored": 0,
+    "fill_rate": 1,
+    "pick_method": "forced",
+    "scaffold_id": "news-briefing",
+    "theme_id": "editorial-navy"
+  },
+  "atomQuality": {
+    "atom_instances": 24,
+    "atom_types_distinct": 13,
+    "variety_ratio": 0.5416666666666666,
+    "unknown_atom_count": 0,
+    "unknown_atom_types": [],
+    "arg_violations": 0,
+    "arg_violations_top_offenders": []
+  },
+  "threeDReadiness": {
+    "twin_coverage": 1,
+    "untwinned_types": [],
+    "lift_success": 14,
+    "lift_attempted": 14,
+    "lift_errors": [],
+    "lifted_subject_rate": 1
+  },
+  "textBudget": {
+    "chars_per_slot": 76.5,
+    "max_slot_chars": 128
+  },
+  "fidelity": {
+    "numbers_total": 18,
+    "numbers_found": 18,
+    "number_recall": 1,
+    "missing_numbers": [],
+    "entities_total": 0,
+    "entities_found": 0,
+    "entity_recall": 1,
+    "missing_entities": []
+  },
+  "score": {
+    "total": 100,
+    "breakdown": {
+      "structure": 25,
+      "variety": 15,
+      "hallucination": 15,
+      "twin": 20,
+      "lift": 15,
+      "text": 10
+    }
+  },
+  "generatedAt": "2026-07-07T01:20:05.709Z"
+}
