+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-01-agenda.json
@@ -1,0 +1,114 @@
+{
+  "slotIdx": 1,
+  "slotName": "agenda",
+  "slotTitle": "Overview",
+  "slotPurpose": "One line per major theme of the briefing",
+  "sourceSlideIdx": 1,
+  "sourceSlideTitle": "核心议题概览",
+  "sceneData": {
+    "name": "news-briefing/agenda",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "核心议题概览",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "agenda-list",
+        "x": 80,
+        "y": 160,
+        "w": 560,
+        "h": 500,
+        "args": {
+          "items": [
+            {
+              "label": "下乡车型销量增长与市场渗透现状"
+            },
+            {
+              "label": "充电桩等配套设施建设进展"
+            },
+            {
+              "label": "主要车企与产品市场格局"
+            },
+            {
+              "label": "农村电网、残值体系等关键挑战"
+            },
+            {
+              "label": "政策退坡对未来增速的影响"
+            }
+          ],
+          "numbered": true
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 680,
+        "y": 200,
+        "w": 520,
+        "h": 180,
+        "args": {
+          "value": ">15%",
+          "label": "五菱宏光MINI单车型贡献",
+          "sublabel": "微型电动车市场集中度",
+          "icon": "car-simple",
+          "style": "accent-border"
+        }
+      },
+      {
+        "type": "bullet-list",
+        "x": 680,
+        "y": 420,
+        "w": 520,
+        "h": 240,
+        "args": {
+          "items": [
+            {
+              "icon": "car-simple",
+              "label": "小型化产品契合农村需求"
+            },
+            {
+              "icon": "currency-dollar-simple",
+              "label": "低价化策略渗透下沉市场"
+            },
+            {
+              "icon": "chart-pie-slice",
+              "label": "细分市场集中度显著"
+            }
+          ]
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "agenda-list",
+      "number-list"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4588,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 623,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.027011399999999998,
+    "elapsedSec": 10.3,
+    "generatedAt": "2026-07-07T01:18:32.120Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
