+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-02-key-figures.json
@@ -1,0 +1,94 @@
+{
+  "slotIdx": 2,
+  "slotName": "key-figures",
+  "slotTitle": "Key Figures",
+  "slotPurpose": "The headline numbers at a glance",
+  "sourceSlideIdx": 2,
+  "sourceSlideTitle": "下乡车型销量高速增长",
+  "sceneData": {
+    "name": "news-briefing/key-figures",
+    "layout": "grid",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "Key Figures",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 80,
+        "y": 180,
+        "w": 340,
+        "h": 420,
+        "args": {
+          "value": "187万辆",
+          "label": "下乡车型销量",
+          "sublabel": "今年前五个月",
+          "icon": "car",
+          "style": "light"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 470,
+        "y": 180,
+        "w": 340,
+        "h": 420,
+        "args": {
+          "value": "+61%",
+          "label": "同比增长",
+          "trend": "up",
+          "trendValue": "+61%",
+          "icon": "chart-line-up",
+          "style": "light"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 860,
+        "y": 180,
+        "w": 340,
+        "h": 420,
+        "args": {
+          "value": "34%",
+          "label": "占全国总销量",
+          "sublabel": "新能源汽车",
+          "icon": "chart-pie",
+          "style": "light"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "stat-grid-large",
+      "kpi-card",
+      "dashboard-multi-kpi-composite"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4446,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 459,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.024125399999999998,
+    "elapsedSec": 7,
+    "generatedAt": "2026-07-07T01:18:39.090Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
