+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-03-theme-1-lead.json
@@ -1,0 +1,87 @@
+{
+  "slotIdx": 3,
+  "slotName": "theme-1-lead",
+  "slotTitle": "Theme 1 — Lead",
+  "slotPurpose": "First major theme: the core facts and figures",
+  "sourceSlideIdx": 3,
+  "sourceSlideTitle": "县乡市场渗透率提升明显",
+  "sceneData": {
+    "name": "news-briefing/theme-1-lead",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "县乡市场渗透率提升明显",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "bar",
+        "x": 80,
+        "y": 180,
+        "w": 1120,
+        "h": 460,
+        "args": {
+          "values": [
+            28,
+            54
+          ],
+          "labels": [
+            "县乡市场",
+            "一二线城市"
+          ],
+          "format": "percent",
+          "title": "市场渗透率对比（当前）",
+          "max": 100
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 920,
+        "y": 180,
+        "w": 280,
+        "h": 140,
+        "args": {
+          "value": "+11%",
+          "label": "县乡市场年增长",
+          "sublabel": "17% → 28%",
+          "trend": "up",
+          "trendValue": "同比去年",
+          "style": "accent-border"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "bar",
+      "column",
+      "stat-banner",
+      "kpi-card"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4459,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 376,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0229194,
+    "elapsedSec": 6.4,
+    "generatedAt": "2026-07-07T01:18:45.537Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
