+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-04-theme-1-detail.json
@@ -1,0 +1,109 @@
+{
+  "slotIdx": 4,
+  "slotName": "theme-1-detail",
+  "slotTitle": "Theme 1 — Detail",
+  "slotPurpose": "First theme drill-down: regional / per-entity breakdown",
+  "sourceSlideIdx": 9,
+  "sourceSlideTitle": "5万至10万元价格带主导市场",
+  "sceneData": {
+    "name": "news-briefing/theme-1-detail",
+    "layout": "grid",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "价格带分布",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "donut-with-center",
+        "x": 80,
+        "y": 180,
+        "w": 520,
+        "h": 480,
+        "args": {
+          "centerValue": "63%",
+          "centerLabel": "5万至10万元",
+          "segments": [
+            {
+              "label": "5万至10万元",
+              "value": 63,
+              "color": [
+                38,
+                70,
+                130
+              ]
+            },
+            {
+              "label": "其他价格带",
+              "value": 37,
+              "color": [
+                200,
+                195,
+                180
+              ]
+            }
+          ],
+          "showPct": true
+        }
+      },
+      {
+        "type": "bullet-list",
+        "x": 680,
+        "y": 180,
+        "w": 520,
+        "h": 480,
+        "args": {
+          "items": [
+            {
+              "icon": "currency-cny",
+              "label": "县乡消费者价格敏感",
+              "status": "highlight"
+            },
+            {
+              "icon": "car-simple",
+              "label": "经济型车型主要受益",
+              "sublabel": "下乡政策带动增长"
+            },
+            {
+              "icon": "chart-line-up",
+              "label": "中低价位市场主导地位"
+            }
+          ]
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "bar",
+      "pie",
+      "segmented-bar",
+      "isotype-stat-comparison"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4458,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 496,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0247164,
+    "elapsedSec": 8.6,
+    "generatedAt": "2026-07-07T01:18:54.147Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
