+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-05-theme-2-lead.json
@@ -1,0 +1,94 @@
+{
+  "slotIdx": 5,
+  "slotName": "theme-2-lead",
+  "slotTitle": "Theme 2 — Lead",
+  "slotPurpose": "Second major theme: the core facts",
+  "sourceSlideIdx": 4,
+  "sourceSlideTitle": "充电桩建设成最大短板",
+  "sceneData": {
+    "name": "news-briefing/theme-2-lead",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "充电桩建设成最大短板",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 80,
+        "y": 200,
+        "w": 340,
+        "h": 400,
+        "args": {
+          "value": "98万台",
+          "label": "县域公共充电桩保有量",
+          "sublabel": "截至5月底全国",
+          "icon": "plugs-connected",
+          "style": "dark"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 470,
+        "y": 200,
+        "w": 340,
+        "h": 400,
+        "args": {
+          "value": "21%",
+          "label": "占全国总量",
+          "sublabel": "县域充电桩占比",
+          "icon": "chart-pie-slice",
+          "style": "dark"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 860,
+        "y": 200,
+        "w": 340,
+        "h": 400,
+        "args": {
+          "value": "<350台",
+          "label": "平均每县",
+          "sublabel": "充电桩数量",
+          "icon": "map-pin",
+          "style": "dark"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "process-arrows",
+      "timeline",
+      "stat-banner",
+      "kpi-card"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4445,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 472,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0243174,
+    "elapsedSec": 7.6,
+    "generatedAt": "2026-07-07T01:19:01.787Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
