+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-07-theme-3-lead.json
@@ -1,0 +1,92 @@
+{
+  "slotIdx": 7,
+  "slotName": "theme-3-lead",
+  "slotTitle": "Theme 3 — Lead",
+  "slotPurpose": "Third major theme: the core facts",
+  "sourceSlideIdx": 7,
+  "sourceSlideTitle": "比亚迪、吉利、长安领跑下乡市场",
+  "sceneData": {
+    "name": "news-briefing/theme-3-lead",
+    "layout": "row",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "市场领跑者",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "kpi-card",
+        "x": 80,
+        "y": 200,
+        "w": 520,
+        "h": 360,
+        "args": {
+          "value": "58%",
+          "label": "市场份额",
+          "sublabel": "比亚迪、吉利、长安合计",
+          "icon": "chart-pie",
+          "style": "dark"
+        }
+      },
+      {
+        "type": "bullet-list",
+        "x": 680,
+        "y": 200,
+        "w": 520,
+        "h": 360,
+        "args": {
+          "items": [
+            {
+              "icon": "trophy",
+              "label": "比亚迪、吉利、长安领跑",
+              "sublabel": "三家车企主导下乡市场"
+            },
+            {
+              "icon": "buildings",
+              "label": "头部车企集中优势",
+              "sublabel": "县乡市场形成强势地位"
+            },
+            {
+              "icon": "storefront",
+              "label": "品牌与渠道下沉",
+              "sublabel": "认知度与渠道能力关键"
+            }
+          ]
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "stat-banner",
+      "kpi-card",
+      "bar",
+      "bullet-list"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4481,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 460,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0242454,
+    "elapsedSec": 8.1,
+    "generatedAt": "2026-07-07T01:19:21.457Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
