+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-08-risk-matrix.json
@@ -1,0 +1,85 @@
+{
+  "slotIdx": 8,
+  "slotName": "risk-matrix",
+  "slotTitle": "Risk Matrix",
+  "slotPurpose": "Risks / scenarios on probability × impact axes",
+  "sourceSlideIdx": 10,
+  "sourceSlideTitle": "挑战一：农村电网改造滞后",
+  "sceneData": {
+    "name": "news-briefing/risk-matrix",
+    "layout": "stage",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "Risk Matrix",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "risk-heatmap",
+        "x": 80,
+        "y": 160,
+        "w": 1120,
+        "h": 520,
+        "args": {
+          "title": "挑战一：农村电网改造滞后",
+          "xAxis": "发生概率",
+          "yAxis": "影响程度",
+          "xLabels": [
+            "很低",
+            "低",
+            "中",
+            "高",
+            "很高"
+          ],
+          "yLabels": [
+            "很低",
+            "低",
+            "中",
+            "高",
+            "很高"
+          ],
+          "risks": [
+            {
+              "x": 3,
+              "y": 2,
+              "label": "部分乡镇变压器容量不足",
+              "description": "夏季用电高峰期充电功率受限"
+            }
+          ]
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "risk-heatmap",
+      "matrix-grid",
+      "swot"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4438,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 342,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0223464,
+    "elapsedSec": 7.3,
+    "generatedAt": "2026-07-07T01:19:28.808Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
