+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-11-quote.json
@@ -1,0 +1,62 @@
+{
+  "slotIdx": 11,
+  "slotName": "quote",
+  "slotTitle": "Key Quote",
+  "slotPurpose": "A load-bearing quote or judgment from the article",
+  "sourceSlideIdx": 6,
+  "sourceSlideTitle": "充电基础设施财政支持力度",
+  "sceneData": {
+    "name": "news-briefing/quote",
+    "layout": "stage",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "Key Quote",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "pull-quote-banner",
+        "x": 80,
+        "y": 180,
+        "w": 1120,
+        "h": 480,
+        "args": {
+          "quote": "中央财政安排专项补贴120亿元，地方配套资金约200亿元，合计投入超过300亿元用于基础设施建设",
+          "author": "充电基础设施财政支持",
+          "bg": "accent"
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "quote-pull",
+      "pull-quote-banner"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4430,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 252,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0209724,
+    "elapsedSec": 5.3,
+    "generatedAt": "2026-07-07T01:19:50.334Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
