+++ sdf-js/examples/scaffold-pipeline/eval-news-policy/slots/slot-12-outlook.json
@@ -1,0 +1,75 @@
+{
+  "slotIdx": 12,
+  "slotName": "outlook",
+  "slotTitle": "Outlook",
+  "slotPurpose": "What to watch next — timeline or conditions",
+  "sourceSlideIdx": 13,
+  "sourceSlideTitle": "前景与展望",
+  "sceneData": {
+    "name": "news-briefing/outlook",
+    "layout": "stage",
+    "subjects": [
+      {
+        "type": "cover",
+        "x": 0,
+        "y": 0,
+        "w": 1280,
+        "h": 120,
+        "args": {
+          "title": "前景与展望",
+          "style": "gradient"
+        }
+      },
+      {
+        "type": "mountain-path",
+        "x": 80,
+        "y": 160,
+        "w": 1120,
+        "h": 520,
+        "args": {
+          "summit": "渗透率提升空间仍然广阔",
+          "milestones": [
+            {
+              "label": "县乡市场增长新引擎",
+              "sublabel": "下沉市场成为增量来源"
+            },
+            {
+              "label": "基础设施配套加速",
+              "sublabel": "电网、充电桩、残值体系"
+            },
+            {
+              "label": "政策市场化平衡",
+              "sublabel": "延续性与转型并重"
+            }
+          ]
+        }
+      }
+    ]
+  },
+  "meta": {
+    "scaffold": "news-briefing",
+    "theme": "editorial-navy",
+    "recommendedAtoms": [
+      "timeline",
+      "mountain-path",
+      "bullet-list"
+    ],
+    "forbiddenAtoms": [],
+    "tokenUsage": {
+      "input_tokens": 4458,
+      "cache_creation_input_tokens": 0,
+      "cache_read_input_tokens": 13008,
+      "cache_creation": {
+        "ephemeral_5m_input_tokens": 0,
+        "ephemeral_1h_input_tokens": 0
+      },
+      "output_tokens": 328,
+      "service_tier": "standard",
+      "inference_geo": "not_available"
+    },
+    "costUSD": 0.0221964,
+    "elapsedSec": 6.9,
+    "generatedAt": "2026-07-07T01:19:57.226Z",
+    "model": "claude-sonnet-4-5-20250929"
+  }
+}
