+++ sdf-js/src/present/scaffolds/mapper-llm.js
@@ -128,6 +128,7 @@ export async function mapSlidesToSlotsLLM(input, opts = {}) {
         'Content-Type': 'application/json',
         'x-api-key': apiKey,
         'anthropic-version': '2023-06-01',
+        'anthropic-dangerous-direct-browser-access': 'true',
       },
       body: JSON.stringify({
         model,
